kubernetes: move k8s-components creation to internal
The CLI will have to create similar objects for k8s upgrades.
